requirements.txt
aioredis==1.3.1
amqp==5.0.6
asgiref==3.3.1
async-timeout==3.0.1
attrs==21.2.0
autobahn==21.3.1
Automat==20.2.0
beautifulsoup4==4.9.3
billiard==3.6.4.0
branca==0.5.0
CacheControl==0.12.10
cachetools==5.0.0
celery==5.1.2
certifi==2020.12.5
cffi==1.14.5
channels==3.0.4
channels-redis==3.3.0
chardet==4.0.0
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.2.0
constantly==15.1.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.4.6
cssutils==2.2.0
daphne==3.0.2
decorator==5.1.1
defusedxml==0.7.1
diff-match-patch==20200713
dj-database-url==0.5.0
Django==3.1.7
django-allauth==0.44.0
django-cors-headers==3.7.0
django-crispy-forms==1.11.1
django-dotenv==1.4.2
django-environ==0.4.5
django-filter==2.4.0
django-import-export==2.5.0
django-inlinecss==0.3.0
django-rest-auth==0.9.5
django-rest-passwordreset==1.1.0
django-timezone-field==4.2.1
djangorestframework==3.12.2
drf-yasg==1.20.0
et-xmlfile==1.1.0
fcm-django==1.0.8
firebase-admin==5.2.0
folium==0.12.1.post1
future==0.18.2
geocoder==1.38.1
google-api-core==2.5.0
google-api-python-client==2.37.0
google-auth==2.6.0
google-auth-httplib2==0.1.0
google-cloud-core==2.2.2
google-cloud-firestore==2.3.4
google-cloud-storage==2.1.0
google-crc32c==1.3.0
google-resumable-media==2.2.1
googleapis-common-protos==1.54.0
grpcio==1.43.0
grpcio-status==1.43.0
gunicorn==20.0.4
hiredis==2.0.0
httplib2==0.20.4
hyperlink==21.0.0
idna==2.10
incremental==21.3.0
inflection==0.5.1
itypes==1.2.0
Jinja2==2.11.3
kombu==5.1.0
MarkupPy==1.14
MarkupSafe==1.1.1
msgpack==1.0.2
numpy==1.22.3
oauthlib==3.1.0
odfpy==1.4.1
openpyxl==3.0.7
packaging==21.0
Pillow==8.1.2
prompt-toolkit==3.0.19
proto-plus==1.20.1
protobuf==3.19.4
psycopg2==2.8.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyJWT==2.0.1
pynliner==0.8.0
pyOpenSSL==20.0.1
pyparsing==2.4.7
python-crontab==2.5.1
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2021.1
PyYAML==5.4.1
ratelim==0.1.6
redis==3.5.3
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.8
ruamel.yaml==0.16.13
ruamel.yaml.clib==0.2.2
service-identity==21.1.0
six==1.15.0
soupsieve==2.2
sqlparse==0.4.1
tablib==3.0.0
Twisted==21.7.0
txaio==21.2.1
typing-extensions==3.10.0.0
uritemplate==3.0.1
urllib3==1.26.3
vine==5.0.0
wcwidth==0.2.5
whitenoise==5.2.0
xlrd==2.0.1
xlwt==1.3.0
zipp==3.4.1
zope.interface==5.4.0
